fix(record): second review pass — races, ownership, and blocking
The reviewer's second pass confirmed eight fixes and found more, including
three my own fixes had introduced. All of them:

Races
- `handOffToRecorder` returned as soon as the child was spawned, but the
  child takes the recording lock later — so `omasnap fullscreen --record;
  omasnap --record --stop` reported nothing running and then left one
  running. The handoff now waits (briefly) for that exact pid to hold the
  lock. Verified with no sleep between the two commands.
- Stopping now signals through a pidfd where the kernel has one, so there is
  no window at all between deciding a pid is our recorder and signalling it;
  it also checks the process is the same user. kill() remains the fallback.
- Pause requests are serialized. State only moved when a reply landed, so a
  second click sent a duplicate whose reply applied the same transition
  twice and double-counted the paused interval.
- The stock-recorder scan moved ahead of crash recovery: an encoder we did
  not spawn may be writing one of those .part.mkv files right now.

Files that are not ours
- The master and the remux draft are created with O_EXCL|O_NOFOLLOW at 0600,
  so a same-named leftover or a planted symlink can no longer be truncated
  by us or by ffmpeg's -y, and the name is taken before the encoder runs.
- Recovery skips symlinks rather than renaming through one and then
  tightening permissions on whatever it pointed at.
- A failed recording no longer renames its master at all: the encoder is
  being shut down and may still hold it open. It stays a .part.mkv and the
  next recording promotes it after checking it can be read.

Blocking and correctness
- Post-remux validation moved off the GUI thread; it ran ffprobe with
  waitForFinished() from the completion callback while "Saving…" was on
  screen. It now also requires a duration above zero rather than just a
  codec name.
- A stop request that is legitimately slow is no longer cut short: the
  budget went from 20s to 120s, since finalizing a long recording on a slow
  filesystem is slow rather than wedged.
- Stopping before the encoder is up is a cancel, not a failure, and says
  nothing rather than reporting that the recording failed.
- `--record-run` rejects any other option instead of ignoring it.
- `connect()` on a unix socket may answer EAGAIN as well as EINPROGRESS.

Tests
- `make smoke` runs ctest rather than named binaries, so switching
  OMASNAP_STUDIO off cannot leave a stale test binary running.
- Both smoke binaries bypass the desktop platform theme the way main() does.
  Without it an offscreen run still loads the gtk3 plugin and fails on a
  machine with no display — which is how the reviewer's first run failed.

+++ src/record-indicator.cpp
@@ -208,9 +208,9 @@ void RecordIndicator::paintEvent(QPaintEvent *) {
   if (phase_ == Phase::Recording) {
     const qreal wave =
         0.5 + 0.5 * std::cos(2.0 * M_PI * pulseTick_ / kPulsePeriod);
-    dot.setAlphaF(0.55 + 0.45 * wave);
+    dot.setAlphaF(static_cast<float>(0.55 + 0.45 * wave));
   } else if (phase_ != Phase::Paused && phase_ != Phase::Failed) {
-    dot.setAlphaF(0.4);
+    dot.setAlphaF(0.4F);
   }
   painter.setPen(Qt::NoPen);
   painter.setBrush(dot);
